requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
django-environ==0.11.2
django-allauth==0.61.1
django-filter==23.5
django-mptt==0.16.0
django-tailwind==3.8.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
djangorestframework==3.14.0
stripe==7.11.0
celery==5.3.6
redis==5.0.1
pillow==10.2.0
gunicorn==21.2.0
python-dateutil==2.8.2
